# Read and decompress columns lazily