… tsl/test/expected/transparent_decompression-14.out
… tsl/test/expected/transparent_decompression-15.out
… tsl/test/shared/expected/transparent_decompress_chunk-14.out
… tsl/test/shared/expected/transparent_decompress_chunk-15.out