We will decompress the compressed columns on demand, skipping them if the vectorized quals don't pass for the entire batch. This allows us to avoid reading some columns, saving on IO. The number of batches that are entirely filtered out is reflected in EXPLAIN ANALYZE as 'Batches Removed by Filters'.

Some benchmarks:
Before, ms | After, ms | Query
--- | --- | ---
73.097 | 37.005 | `select distinct(country) from uk_price_paid where price > 150000000;`
19.149 | 11.295 | `select count(*), avg(total_amount), max(dropoff_datetime) from rides where passenger_count >= 5;`


Disable-check: force-changelog-file